% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binarydb.R
\name{binarydb}
\alias{binarydb}
\alias{binarydb_down}
\alias{binarydb_load}
\title{CRAN Matrix Of Available Binary Packages (archivedb.rda)}
\usage{
binarydb_down(dir = ".", repos = getOption("repos")[1])

binarydb_load(filename = "binarydb.rda")
}
\arguments{
\item{dir}{character. The directory where "binarydb.rda" is saved
Default value \code{"."} is the current directory.}

\item{repos}{character. The address of your local CRAN.}

\item{filename}{character. The (path to a) file "binarydb.rda" or an equivalent.}
}
\description{
\code{binarydb_down} downloads the matrix of the packages available in binary
format available in CRAN (Windows or macOS depending the computer), saves it
it in a file named \code{binarydb} and finally loads it in .GlobalEnv under
the name \code{binarydb}. It is a wrapper around the function
\code{available.packages(type = "binary")}.

\code{binarydb_load} loads the file \code{filename} in .GlobalEnv under the
name \code{binarydb}.

When loaded in .GlobalEnv, \code{binarydb} is recognized by the functions
\code{\link{p_vers}} and \code{\link{p_vers_deps}} to compare the package
version numbers of the installed packages with the most recent versions of
the binary and the source versions available in CRAN.
}
\examples{
## In real life, to get the latest version, use
## binarydb_down()
## Here, we use a small file
binarydb_load(system.file("data", "zbinarydb.rda", package = "RWsearch"))
class(binarydb)
colnames(binarydb)
dim(binarydb)
head(binarydb)

}
