% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_pdf.R
\name{f_pdf}
\alias{f_pdf}
\title{PDF Pages of Functions}
\usage{
f_pdf(..., char = NULL)
}
\arguments{
\item{...}{any format recognized by \code{\link{cnsc}}, except list.
A vector of quoted "package::function".}

\item{char}{(name to) a character vector. Use this argument if 
\code{...} fails or if you call the function from another function. 
If used, argument \code{...} is ignored.}
}
\description{
\code{f_pdf} generates in the current directory the pdf pages of one or several functions. 
The pdf pages are printed but not opened. 
Miktex or Texlive is required. See also packages \emph{document} and \emph{sinew}.
}
\examples{
\donttest{
f_pdf(c("RWsearch::cnsc", "RWsearch::p_inun"))
}
}
