% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_engine.R
\name{h_engine}
\alias{h_engine}
\alias{h_1bis}
\alias{h_abcbourse}
\alias{h_ask}
\alias{h_arxiv}
\alias{h_arxivpdf}
\alias{h_baidu}
\alias{h_blackle}
\alias{h_bing}
\alias{h_biorxiv}
\alias{h_biorxivpdf}
\alias{h_bmap}
\alias{h_boursorama}
\alias{h_cnrtl}
\alias{h_collins}
\alias{h_cpan}
\alias{h_crossref}
\alias{h_ctan}
\alias{h_daum}
\alias{h_ddg}
\alias{h_dm}
\alias{h_doi}
\alias{h_ecosia}
\alias{h_egerin}
\alias{h_estrep}
\alias{h_evene}
\alias{h_exalead}
\alias{h_excite}
\alias{h_framabee}
\alias{h_gigablast}
\alias{h_github}
\alias{h_gitlab}
\alias{h_gmap}
\alias{h_google}
\alias{h_gscholar}
\alias{h_info}
\alias{h_ixquick}
\alias{h_khoj}
\alias{h_lesechos}
\alias{h_lilo}
\alias{h_lt}
\alias{h_lycos}
\alias{h_mappy}
\alias{h_mw}
\alias{h_nate}
\alias{h_naver}
\alias{h_orcid}
\alias{h_osm}
\alias{h_osmn}
\alias{h_parsijoo}
\alias{h_peertube}
\alias{h_peru}
\alias{h_pipilika}
\alias{h_qwant}
\alias{h_qwfr}
\alias{h_reverso_d}
\alias{h_sapo}
\alias{h_searx}
\alias{h_so}
\alias{h_sogou}
\alias{h_ssrn}
\alias{h_ssrnauth}
\alias{h_startpage}
\alias{h_twfr}
\alias{h_twitter}
\alias{h_un}
\alias{h_verbes}
\alias{h_via}
\alias{h_vimeo}
\alias{h_wego}
\alias{h_wp}
\alias{h_wpfr}
\alias{h_yahoo}
\alias{h_yahoofin}
\alias{h_yandex}
\alias{h_yooz}
\alias{h_yt}
\title{Explore the Web with Various Search Engines}
\usage{
h_1bis(..., char = NULL)

h_abcbourse(..., char = NULL)

h_ask(..., char = NULL)

h_arxiv(..., char = NULL)

h_arxivpdf(..., char = NULL)

h_baidu(..., char = NULL)

h_blackle(..., char = NULL)

h_bing(..., char = NULL)

h_biorxiv(..., char = NULL)

h_biorxivpdf(..., char = NULL)

h_bmap(..., char = NULL)

h_boursorama(..., char = NULL)

h_cnrtl(..., char = NULL)

h_collins(..., char = NULL)

h_cpan(..., char = NULL)

h_crossref(..., char = NULL)

h_ctan(..., char = NULL)

h_daum(..., char = NULL)

h_ddg(..., char = NULL)

h_dm(..., char = NULL)

h_doi(..., char = NULL)

h_ecosia(..., char = NULL)

h_egerin(..., char = NULL)

h_estrep(..., char = NULL)

h_evene(..., char = NULL)

h_exalead(..., char = NULL)

h_excite(..., char = NULL)

h_framabee(..., char = NULL)

h_gigablast(..., char = NULL)

h_github(..., char = NULL)

h_gitlab(..., char = NULL)

h_gmap(..., char = NULL)

h_google(..., char = NULL)

h_gscholar(..., char = NULL)

h_info(..., char = NULL)

h_ixquick(..., char = NULL)

h_khoj(..., char = NULL)

h_lesechos(..., char = NULL)

h_lilo(..., char = NULL)

h_lt(..., char = NULL)

h_lycos(..., char = NULL)

h_mappy(..., char = NULL)

h_mw(..., char = NULL)

h_nate(..., char = NULL)

h_naver(..., char = NULL)

h_orcid(..., char = NULL)

h_osm(..., char = NULL)

h_osmn(..., char = NULL)

h_parsijoo(..., char = NULL)

h_peertube(..., char = NULL)

h_peru(..., char = NULL)

h_pipilika(..., char = NULL)

h_qwant(..., char = NULL, lang = "en")

h_qwfr(..., char = NULL)

h_reverso_d(..., char = NULL)

h_sapo(..., char = NULL)

h_searx(..., char = NULL)

h_so(..., char = NULL)

h_sogou(..., char = NULL)

h_ssrn(..., char = NULL)

h_ssrnauth(..., char = NULL)

h_startpage(..., char = NULL)

h_twfr(..., char = NULL)

h_twitter(..., char = NULL, lang = "en")

h_un(..., char = NULL)

h_verbes(..., char = NULL)

h_via(..., char = NULL)

h_vimeo(..., char = NULL)

h_wego(..., char = NULL)

h_wp(..., char = NULL, lang = "en")

h_wpfr(..., char = NULL)

h_yahoo(..., char = NULL, lang = "en")

h_yahoofin(..., char = NULL, lang = "en")

h_yandex(..., char = NULL)

h_yooz(..., char = NULL)

h_yt(..., char = NULL)
}
\arguments{
\item{...}{any format recognized by \code{\link{cnsc}}, except list.
A vector of packages.}

\item{char}{(name to) a character vector. Use this argument if 
\code{...} fails or if you call the function from another function.}

\item{lang}{character. The language accepted by the search engine, usually 
"en", "de", "es", "fr", "jp", etc.}
}
\description{
Launch the default browser and search in: 1bis Map (BottinCarto), ABC Bourse (short 
stock names), arXiv (vectorized), Ask, Baidu, Blackle, Bing, Bing Map (bmap), 
Boursorama (short stocknames), CNRTL (French dictionary), 
Collins English Dictionary, CPAN and metaCPAN (Perl), 
Crossref (DOI and bibliographic metadata), CTAN (Latex), Daum, DailyMotion (dm), 
DOI, DuckDuckGo (ddg), Ecosia, Egerin, Evene (citations), Exalead, Excite, 
Gigablast, GitHub, GitLab, Google Map (gmap), Google, Google Scholar (gscholar), Info, 
Khoj, Les Echos, La Tribune (lt), Lilo, Lycos, Mappy Map, Merriam-Webster (mw, English dictionary), 
Nabble, Nate, Naver (see N2H4 package), Orcid, Open Street Map, OSM Nominatim, 
Parsijoo, PeerTube, Peru, Pipilika, 
Qwant (qw + qwfr), R-bloggers, Rdocumentation (rdoc), Rdocumentation task views (rdoctv), 
Rdrr, Reverso dictionnary, Rseek, Sapo, Searx, Sogou, SSRN and SSRN Author (vectorized),
Stackoverflow (so), Startpage (ex-Ixquick), Twitter (+ twfr), 
L'Usine Nouvelle (un), ViaMichelin Map and Routes (via), Les Verbes, Vimeo, Wego (Here maps), 
Wikipedia (wp + wpfr), Yahoo, Yahoo Finance, Yandex, Yooz, Youtube (yt). 

Using the regular R format "w1 w2 w3" rather than w1, w2, w3 makes sense as most
functions collapse the words into character chains "w1 w2 w3", "w1+w2+w3" or "w1-w2-w3".

Visit \url{https://en.wikipedia.org/wiki/Web_search_engine} for a list of web search
engines.
}
\examples{
\donttest{
h_yt("Serge Gainsbourg Ne dis rien")
h_so(R, deep, neural, network)
h_osm("Le Chateau d'Oleron")
h_mw(recension)
h_arxiv(c(1212.4320, 1605.08732))
h_doi("10.1016/j.ejor.2013.06.029")
}
}
