% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_R.R
\name{h_R}
\alias{h_R}
\alias{h_cran}
\alias{h_cranbydate}
\alias{h_cranbyname}
\alias{h_crantv}
\alias{h_cranstatus}
\alias{h_cranberries}
\alias{h_gepuro}
\alias{h_nabble}
\alias{h_rbloggers}
\alias{h_rdoc}
\alias{h_rdoctv}
\alias{h_rdrr}
\alias{h_rseek}
\title{Open a Web Page in the Browser}
\usage{
h_R()

h_cran(repos = getOption("repos")[1])

h_cranbydate(repos = getOption("repos")[1])

h_cranbyname(repos = getOption("repos")[1])

h_crantv(repos = getOption("repos")[1])

h_cranstatus()

h_cranberries()

h_gepuro()

h_nabble(..., char = NULL)

h_rbloggers(..., char = NULL)

h_rdoc(..., char = NULL)

h_rdoctv(..., char = NULL)

h_rdrr(..., char = NULL)

h_rseek(..., char = NULL)
}
\arguments{
\item{repos}{character. The address of your local CRAN.}

\item{...}{any format recognized by \code{\link{cnsc}}, except list.
A regular web address.}

\item{char}{(name to) a character vector. Use this argument if 
\code{...} fails or if you call the function from another function.}
}
\description{
\code{h_R} opens the page \url{https://www.r-project.org}.

\code{h_cran} opens the page of you local CRAN.  

\code{h_crandate} opens the page of CRAN packages sorted by date of publication.

\code{h_crantv} opens the page of CRAN task views.

\code{h_cranberries}, \code{h_nabble}, \code{h_rbloggers}, \code{h_rdoc}, \code{h_rdoctv} 
(RDocumentation), \code{h_rdrr}, \code{h_rseek} open the pages of web sites related 
to R.

\code{h_gepuro} lists all (most) R packages available on GitHub.
}
\examples{
\donttest{
h_cranbydate()
h_cranberries()
}
}
