% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_ttp.R
\name{h_ttp}
\alias{h_ttp}
\title{Open a Web Page in the Browser}
\usage{
h_ttp(..., char = NULL, https = TRUE, www = FALSE)
}
\arguments{
\item{...}{any format recognized by \code{\link{cnsc}}, except list.
A regular web address.}

\item{char}{(name to) a character vector. Use this argument if 
\code{...} fails or if you call the function from another function.}

\item{https}{logical. Use https or http.}

\item{www}{logical. Add www. to the address.}
}
\description{
\code{h_ttp} opens the page coresponding to the mentionned address in the default browser.
}
\examples{
\donttest{
h_ttp("www.r-project.org")
}
}
