% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_crandb_tvdb.R
\name{s_crandb_tvdb}
\alias{s_crandb_tvdb}
\title{Search For Recent Packages In crandb And In Task View}
\usage{
s_crandb_tvdb(..., char = NULL, tv = "Distributions", from = -10,
  to = Sys.Date(), select = "PTD", mode = "or", sensitive = FALSE,
  perl = FALSE, fixed = FALSE, agrep = FALSE, max.distance = 0.1,
  costs = NULL, crandb = get("crandb", envir = .GlobalEnv),
  tvdb = get("tvdb", envir = .GlobalEnv))
}
\arguments{
\item{...}{any format recognized by \code{\link{cnsc}}, except list.
One or several keywords.}

\item{char}{(name to) a character vector. Use this argument if \code{...} fails 
or if you call the function from another function. If used, 
argument \code{...} is ignored.}

\item{tv}{character. One task view among those listed in tvdb.}

\item{from}{character representing a date earlier than date \code{to}. 
Or a negative integer representing the number of days preceeding 
the date \code{to}.}

\item{to}{date. The upper date in the search.}

\item{select}{character vector. A sub-vector of \code{colnames(crandb)}. The 
short form "P", "T", "D", "PT", "PD", "TD", "PTD", "A", "M", "AM" 
describing the Package name, Title, Description, Author, Maintainer
or a combination of them is accepted.}

\item{mode}{character among "or", "and", "relax". The search mode. 
"relax" is for 3 words and more. It is an intermediate between 
"or" and "and" as it requires just 2 matching words: 
\code{("word1" AND "word2")} \code{OR} \code{("word1" AND "word3")} 
\code{OR} \code{("word1" AND "word3")}.}

\item{sensitive}{logical. \code{TRUE} forces the search to be case sensitive.}

\item{perl}{logical. Used only if \code{fixed = FALSE}. \code{TRUE} uses 
Perl-compatible regex. \code{FALSE} uses default regexps.}

\item{fixed}{logical. \code{TRUE} matchs the keywords as is (and \code{sensitive} 
is forced to \code{TRUE}). \code{FALSE} allows grep or Perl regexps. 
See \code{\link[base]{grep}}. Not used by \code{agrep}.}

\item{agrep}{logical. For approximate matching, use \code{\link[base]{agrep}} 
function rather than grep.}

\item{max.distance}{integer or numeric. See \code{\link[base]{agrep}}.}

\item{costs}{NULL or list. See \code{\link[base]{agrep}}.}

\item{crandb}{data.frame \code{crandb}.}

\item{tvdb}{list. The list of the task views.}
}
\value{
A list with the following vectors:
\itemize{
  \item{spkgs: }{The selected packages that contain the keyword(s).}
  \item{inTV: }{The packages that contain the keyword(s) already refereed in the task view.}
  \item{notinTV: }{The packages that contain the keyword(s) not (yet) refereed in the task view.}
  \item{inTV_in: }{Among the packages available in the task view, those installed in the computer.}
  \item{inTV_un: }{Among the packages available in the task view, those not installed in the computer.}
  \item{notinTV_in: }{Among the packages not refereed in the task view, those installed in the computer.} 
  \item{notinTV_un: }{Among the packages not refereed in the task view, those not installed in the computer.}
}
}
\description{
This is a function for task view maintenance.

Search packages in a subset of \code{crandb} within dates \code{from} and \code{to} 
that contain one or several keywords in the columns "Package", "Title", "Description", 
'Author" or "Maintainer", then verify if these packages are already refereed in one 
of the task views stored in \code{tvdb}.
}
\examples{
### TASK VIEW MAINTENANCE (tvdb + crandb)
## In real life, download crandb and tvdb from CRAN or load them from your directory 
## with functions crandb_down(), crandb_load(), tvdb_down(), tvdb_load(). 
## In this example, we use small files.
crandb_load(system.file("data", "zcrandb.rda", package = "RWsearch"))
tvdb_load(system.file("data", "ztvdb.rda", package = "RWsearch")) 

## List the task views
tvdb_vec()

## Search for the recent packages in crandb that contain the keyword 
## and verify the packages already refereed in the task view.
## With ztvdb.rda, from = "2018-01-01" is selected.
(lst <- s_crandb_tvdb("distribution", tv = "Distributions", from = "2018-01-01"))
\donttest{
## Display in the browser some information about the selected packages.
p_display7(lst[c("inTV", "notinTV")], dir = tempdir())
}

}
