% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_hs.R
\name{s_hs}
\alias{s_hs}
\title{Search Packages and Functions in Installed Packages}
\usage{
s_hs(..., char = NULL, fields = c("alias", "concept", "title"),
  apropos, keyword, whatis, ignore.case = TRUE, package = NULL,
  agrep = NULL, use_UTF8 = FALSE)
}
\arguments{
\item{...}{one single character string recognized by \code{\link{cnsc}}.
One and only one pattern}

\item{char}{(name to) a single character string. Use this argument if \code{...} 
fails or if you call the function from another function. If used, 
argument \code{...} is ignored.}

\item{fields}{See \code{help.search}.}

\item{apropos}{See \code{help.search}.}

\item{keyword}{See \code{help.search}.}

\item{whatis}{See \code{help.search}.}

\item{ignore.case}{See \code{help.search}.}

\item{package}{See \code{help.search}.}

\item{agrep}{See \code{help.search}.}

\item{use_UTF8}{See \code{help.search}.}
}
\description{
\code{s_hs} is a wrapper of the well known function \code{??} and its parent function
\code{help.search}. Visit the help page \code{\link[utils]{help.search}} for details 
on the various arguments.
}
\examples{
\donttest{ 
s_hs("neural network")
}
}
