% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_sos.R
\name{s_sos}
\alias{s_sos}
\alias{s_man}
\alias{s_universe}
\title{Search Packages and Functions on R-Project Help pages and RDocumentation}
\usage{
s_sos(..., char = NULL)

s_man(..., char = NULL)

s_universe(..., char = NULL)
}
\arguments{
\item{...}{any format recognized by \code{\link{cnsc}}, except list.
One or several keywords.}

\item{char}{(name to) a character vector. Use this argument if \code{...} fails
or if you call the function from another function. If used,
argument \code{...} is ignored.}
}
\description{
\code{s_sos} searchs in all R documentation packages and functions that contain
one or several keywords, open the default browser and display the results in two 
html pages, one for the package::functions and one for the package::vignettes.
\code{s_sos} usually find more results than \code{\link{s_crandb}} as it goes 
deeper in the documentation, down to the function level. An internet connection 
is required to reach the website \url{https://search.r-project.org}. 
\code{s_sos} is a minimal wrapper of the function \emph{sos::findFn}.
Use this function if you need advanced search options.

\code{s_man} sends a query to the same search engine but the results is returned 
in a less edited format than \code{s_sos}. Use the function \code{\link{p_man}} 
if your search is about a package and not a function.

\code{s_universe} sends a query to the r-universe.dev website which encompass
CRAN, Bioconductor and Git repositories.
}
\examples{
## Search using standard or non-standard content
## and display the results in a browser.
if (interactive()) {
s_sos("chemical reaction")
(res <- s_sos(distillation))
tail(data.frame(res))
s_man("cran_incoming")
s_universe(back, testing)
}
}
\seealso{
\url{https://CRAN.R-project.org/package=sos} (index and vignette).
}
