% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_intro.R
\docType{data}
\name{ztvdb}
\alias{ztvdb}
\title{File ztvdb.rda: A Subset of tvdb Dataset}
\description{
File \emph{ztvdb.rda} is a small file of 5 ko that contains 6 task views and acts as 
a replacement of the large file \emph{tvdb.rda} downloaded from CRAN that contains  
36 task views. It loads in .GlobalEnv as \code{tvdb}. The use of \emph{ztvdb.rda} 
avoids inappropriate connections to CRAN and increases the speed in the examples.
}
\examples{
tvdb_load(system.file("data", "ztvdb.rda", package = "RWsearch"))
}
\keyword{datasets}
