% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_check.R
\name{e_check}
\alias{e_check}
\title{Check Results of Packages Identified by their Email Address}
\usage{
e_check(..., char = NULL, repos = getOption("repos")[1])
}
\arguments{
\item{...}{any format recognized by \code{\link{cnsc}}, except list.
A vector of quoted "e-mail adresses".}

\item{char}{(name to) a character vector. Use this argument if
\code{...} fails or if you call the function from another function.
If used, argument \code{...} is ignored.}

\item{repos}{character. The address of your local CRAN.}
}
\description{
\code{e_check} opens the browser and returns the "CRAN Check Results" page(s)
of the packages maintained by one or several maintainers identified by their
regular email addresses (but not the orphaned ones). An internet connection
is required. This function is a simplified version of the functions proposed
in the package \emph{foghorn}.
}
\examples{
if (interactive()) {
e_check(c("rpackages@inmodelia.com", "christophe.dutang@ensimag.fr"),
        repos = "https://cloud.r-project.org")
}
}
