% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_pdf.R
\name{f_pdf}
\alias{f_pdf}
\title{PDF Pages of Functions}
\usage{
f_pdf(..., char = NULL)
}
\arguments{
\item{...}{any format recognized by \code{\link{cnsc}}, except list.
A vector of quoted "package::function".}

\item{char}{(name to) a character vector. Use this argument if
\code{...} fails or if you call the function from another function.
If used, argument \code{...} is ignored.}
}
\description{
\code{f_pdf} generates in the current directory the pdf pages of one or several
functions. The pdf pages are printed but not opened. Miktex or Texlive is required.
This function wraps \code{{utils::help}} with the pdf option activated.
Similar functions are available in the packages \emph{document} and \emph{sinew}.
}
\examples{
## FALSE is here to avoid a NOTE in CRAN checks. Ignore this line.
if (FALSE) {
   f_pdf(c("RWsearch::cnsc", "RWsearch::p_inun"))
}
}
