% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_direct.R
\name{h_direct}
\alias{h_direct}
\alias{h_academie}
\alias{h_framasoft}
\alias{h_framasoft0}
\alias{h_lexilogos}
\alias{h_linguee}
\alias{h_reverso}
\alias{h_tad}
\alias{h_tadsm}
\title{Open a Web Page in the Browser}
\usage{
h_academie()

h_framasoft()

h_framasoft0()

h_lexilogos()

h_linguee()

h_reverso()

h_tad()

h_tadsm()
}
\description{
\code{h_academie} gives access to the Academie francaise's dictionary.

\code{h_framasoft} gives access to several free web services (as in speech and in 
half pint of beer) that are good alternatives to GAFA services. 

\code{h_lexilogos} gives access to hundreds of dictionaries in many languages.

\code{h_linguee} gives access to a translation engine in many languages. Many
examples with long sentences.

\code{h_reverso} gives access to a translation engine and dictionnaries in many languages.

\code{h_tad} and \code{h_tadsm} gives access to a website dedicated to date and time 
conversion plus timezone management.

Using the regular R format "w1 w2 w3" rather than w1, w2, w3 makes sense as most
engines collapse the words into character chains "w1 w2 w3", "w1+w2+w3" or "w1-w2-w3".
}
\examples{
\donttest{
h_linguee()
h_lexilogos()
}
}
