% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_R.R
\name{h_R}
\alias{h_R}
\alias{h_Rblog}
\alias{h_Rman}
\alias{h_Rml}
\alias{h_Rnews}
\alias{h_Rversions}
\alias{h_cran}
\alias{h_cranbydate}
\alias{h_cranbyname}
\alias{h_cranchecks}
\alias{h_crancheckwindows}
\alias{h_crantv}
\alias{h_cranstatus}
\alias{h_cranberries}
\alias{h_gepuro}
\alias{h_nabble}
\alias{h_rbloggers}
\alias{h_rdoc}
\alias{h_rdoctv}
\alias{h_rdrr}
\alias{h_rseek}
\alias{h_biocstats}
\title{Open a Web Page in the Browser}
\usage{
h_R()

h_Rblog()

h_Rman()

h_Rml()

h_Rnews(repos = getOption("repos")[1])

h_Rversions(repos = getOption("repos")[1])

h_cran(repos = getOption("repos")[1])

h_cranbydate(repos = getOption("repos")[1])

h_cranbyname(repos = getOption("repos")[1])

h_cranchecks(repos = getOption("repos")[1])

h_crancheckwindows(repos = getOption("repos")[1])

h_crantv(repos = getOption("repos")[1])

h_cranstatus()

h_cranberries()

h_gepuro()

h_nabble(..., char = NULL)

h_rbloggers(..., char = NULL)

h_rdoc(..., char = NULL)

h_rdoctv(..., char = NULL)

h_rdrr(..., char = NULL)

h_rseek(..., char = NULL)

h_biocstats()
}
\arguments{
\item{repos}{character. The address of your local CRAN.}

\item{...}{any format recognized by \code{\link{cnsc}}, except list.
A regular web address.}

\item{char}{(name to) a character vector. Use this argument if
\code{...} fails or if you call the function from another function.}
}
\description{
\code{h_R} opens the page \url{https://www.r-project.org}.
\code{h_Rml} opens the page dedicated to the mailing lists
\url{https://www.r-project.org/mail.html}.
\code{h_Rnews} opens the page \url{https://cran.r-project.org/doc/manuals/r-devel/NEWS.html}.
\code{h_Rversions} opens a page (from rversions package) that keeps a record
of all R versions and their release dates.

\code{h_cran} opens the page of you local CRAN.

\code{h_cranbydate} and \code{h_cranbyname} open the page of CRAN packages
sorted by date of publication and in alphabetical order.

\code{h_cranchecks} and \code{h_crancheckwindows} open the pages related to
the checks of all packages listed by name, maintainers, dates, os. A special
page is dedicated to Windows packages with the results for the previous, the
current and the devel R versions.

\code{h_crantv} opens the page of CRAN task views.

\code{h_cranberries}, \code{h_nabble}, \code{h_rbloggers}, \code{h_rdoc}, \code{h_rdoctv}
(RDocumentation), \code{h_rdrr}, \code{h_rseek} open the pages of web sites related
to R.

\code{h_gepuro} lists all (most) R packages available on GitHub. A huge file.
}
\examples{
if (interactive()) {
h_crantv(repos = "https://cloud.r-project.org")
h_cranberries()
}
}
