% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_vers.R
\name{p_vers}
\alias{p_vers}
\title{Package Version}
\usage{
p_vers(..., char = NULL)
}
\arguments{
\item{...}{any format recognized by \code{\link{cnsc}}, except list.
A vector of packages.}

\item{char}{(name to) a character vector. Use this argument if \code{...} 
fails or if you call the function from another function. 
If used, argument \code{...} is ignored.}
}
\description{
Return a data.frame with the version number of each package. \code{NA} and 
a warning is issued if the package is not installed. A vectorized version of 
\code{utils::packageVersion}.
}
\examples{
p_vers(RWsearch, MASS, Matrix, survival, R)

}
