% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_sos.R
\name{s_sos}
\alias{s_sos}
\title{Search Packages and Functions in U. of Pennsylvania and RDocumentation}
\usage{
s_sos(..., char = NULL)
}
\arguments{
\item{...}{any format recognized by \code{\link{cnsc}}, except list.
One or several keywords.}

\item{char}{(name to) a character vector. Use this argument if \code{...} fails 
or if you call the function from another function. If used, 
argument \code{...} is ignored.}
}
\description{
\code{s_sos} searchs in all R documentation packages and functions that contain 
one or several keywords, open the default browser and display the results in a html page. 
For one or two keywords, \code{s_sos} may find more results than \code{\link{s_crandb}}  
as it goes deeper in the documentation, down to the function level. An internet 
connection is required to reach the website maintained by the University of Pennsylvania.

\code{s_sos} is a minimal wrapper of the function \emph{sos::findFn}. 
Use directly the package \emph{sos} and read its vignette for advanced search options.
}
\examples{
\donttest{
## Search using standard or non-standard content
## and display the results in a browser. 
s_sos(distillation)

res <- s_sos("chemical reaction")
as.data.frame(res)
res
}
}
\seealso{
\url{http://finzi.psych.upenn.edu/}, 

          \url{https://CRAN.R-project.org/package=sos} (index and vignette).
}
