% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_intro.R
\docType{data}
\name{ztvdb}
\alias{ztvdb}
\title{File ztvdb.rda: A Subset of tvdb Dataset}
\description{
The file \emph{ztvdb.rda} contains a list of 6 task views named \code{tvdb}. 
It is a subset of the large file \emph{tvdb.rda} that contain 42 task views
usually downloaded from CRAN by the function \code{\link{tvdb_down}}.
The use of \emph{ztvdb.rda} avoids inappropriate connections to CRAN and
increases the speed of the examples.
}
\examples{
tvdb_load(system.file("data", "ztvdb.rda", package = "RWsearch"))
}
\keyword{datasets}
